requirements.txt
sklearn-pandas   ==  1.8.0
hyperopt         ==  0.1.2
scikit-learn     ==  0.20.0
openpyxl         ==  2.5.4
PyForms          ==  4.0.3
remi             == 2019.4
statsmodels      ==  0.11.0
numpy             == 1.15.4
xlrd            ==  1.2.0
pillow          == 6.0.0
matplotlib      == 2.2.2
pandas        == 0.25.3
networkx        ==   1.11

